finished most endpoints
